[scsi] qla2xxx: Retain loop test for fwdump length exceeding buffer length
Message-id: <20170927214454.8537-36-hmadhani@redhat.com>
Patchwork-id: 186863
O-Subject: [RHEL 7.5 e-stor PATCH v2 35/96] scsi: qla2xxx: Retain loop test for fwdump length exceeding buffer length
Bugzilla: 1460030
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Joe Carnuccio <joe.carnuccio@cavium.com>

Bugzilla 1460030

commit 383a298b2065a6f39f4ab2c33d84620ccad578f9

[ hmadhani: RH source is different because it is missing patch ]
[           which added chages to qla_init.c. ]

---
 drivers/scsi/qla2xxx/qla_init.c |  4 ++--
 drivers/scsi/qla2xxx/qla_tmpl.c | 16 +++++++++-------
 2 files changed, 11 insertions(+), 9 deletions(-)

+++ drivers/scsi/qla2xxx/qla_init.c
@@ -6374,7 +6374,7 @@ qla24xx_load_risc_flash(scsi_qla_host_t *vha, uint32_t *srisc_addr,
 	if (dlen > risc_size * sizeof(*dcode)) {
 		ql_log(ql_log_warn, vha, 0x0167,
 		    "Failed fwdump template exceeds array by %lx bytes\n",
-		    dlen - risc_size * sizeof(*dcode));
+		    (size_t)(dlen - risc_size * sizeof(*dcode)));
 		goto default_template;
 	}
 	ha->fw_dump_template_len = dlen;
@@ -6676,7 +6676,7 @@ qla24xx_load_risc_blob(scsi_qla_host_t *vha, uint32_t *srisc_addr)
 	if (dlen > risc_size * sizeof(*fwcode)) {
 		ql_log(ql_log_warn, vha, 0x0177,
 		    "Failed fwdump template exceeds array by %lx bytes\n",
-		    dlen - risc_size * sizeof(*fwcode));
+		    (size_t)(dlen - risc_size * sizeof(*fwcode)));
 		goto default_template;
 	}
 	ha->fw_dump_template_len = dlen;
